glemu deadcode and macro purge
Each line was removed, one by one, check that the compilation passed.
On Debian 10, with clang++ and some custum setup, all passes. CI will
check that all damn systems builds. It's the only thing it's setup for,
after all.
I will *not* apply ANY codestyle or doc patch on this, because it's just
a damn cleanup, not real changes.
To redo:
* 1: apply `cpp` on your target
* 2: n = 1
* 3: until EOF
* 4: remove line 1
* 5: build
* 6: if build failed, cancel 3
* 7: n++
* 8: goto 3